Add content type to response

+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/sms/outbound/OutboundSmsService.java
@@ -40,7 +40,7 @@ public interface OutboundSmsService
 
     long saveOutboundSms( OutboundSms sms );
 
-    void deleteById( Integer outboundSmsId );
+    void deleteById( Long outboundSmsId );
 
     void deleteById( String uid );
