+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/sms/outbound/DefaultOutboundSmsService.java
@@ -84,7 +84,7 @@ public class DefaultOutboundSmsService
     }
 
     @Override
-    public void deleteById( Integer outboundSmsId )
+    public void deleteById( Long outboundSmsId )
     {
         OutboundSms sms = outboundSmsStore.getOutboundSmsbyId( outboundSmsId );
